Organize and document.
