+++ lib/rust/ensogl/pack/src/lib.rs
@@ -1,6 +1,6 @@
-//! EnsoGL Pack compiles Rust sources, precompile shaders of EnsoGL app, and outputs JS WASM loader,
-//! additional JS runtime utilities, and a set of optimized shaders. It is a wrapper for `wasm-pack`
-//! tool.
+//! EnsoGL Pack compiles Rust sources, builds the dynamic assets of the EnsoGL app (including
+//! optimized shaders and pre-seeded caches), and outputs the JS WASM loader, additional JS runtime
+//! utilities, and a set of optimized dynamic assets. It is a wrapper for `wasm-pack` tool.
 //!
 //! # Compilation process.
 //! When run, the following file tree will be created/used. The files/directories marked with '*'
@@ -25,13 +25,11 @@
 //!       │  ├─ runtime-libs.js          | Bundled `this_crate/js/runtime-libs`.
 //!       │  ╰─ snippets                 | Rust-extracted JS snippets.
 //!       │     ╰─ <name>.js             | A single Rust-extracted JS snippet.
-//!       ├─ shaders                     | Not optimized shaders sources extracted from WASM bundle.
-//!       │  ├─ list.txt                 | List of extracted not optimized shaders (no extensions).
-//!       │  ├─ <name>.<stage>.glsl      | A single not optimized shader. (Stage = vertex|fragment).
-//!       │  ╰─ ...
-//!       ├─ shaders-hash                | Not optimized shader hashes. Used to detect changes.
-//!       │  ├─ <name>.<stage>.hash      | A single not optimized shader hash.
-//!       │  ╰─ ...
+//!       ├─ dynamic-assets              | Dynamic asset sources extracted from WASM bundle.
+//!       │  ╰─ <type>                   | A subdirectory for each type of asset.
+//!       │     ├─ <key>                 | A subdirectory for each asset's sources.
+//!       │     ├─ <key>.work            | Intermediate files produced when building an asset.
+//!       │     ╰─ ...
 //!       ├─ runtime-libs
 //!       │  ╰─ runtime-libs.js
 //!       ├─ linked-dist                 | Either symlink to dist or to the gui artifacts.
@@ -39,15 +37,17 @@
 //!        * ├─ index.js                 | The main JS bundle to load WASM and JS wasm-pack bundles.
 //!          ├─ index.js.map             | The sourcemap mapping to sources in TypeScript.
 //!       ** ├─ index.d.ts               | TypeScript types interface file.
-//!          ├─ shader-extractor.js      | Node program to extract non optimized shaders from WASM.
-//!          ├─ shader-extractor.js.map  | The sourcemap mapping to sources in TypeScript.
-//!          ├─ shader-extractor.d.ts    | TypeScript types interface file.
+//!          ├─ asset-extractor.js       | Node program to extract asset sources from WASM.
+//!          ├─ asset-extractor.js.map   | The sourcemap mapping to sources in TypeScript.
+//!          ├─ asset-extractor.d.ts     | TypeScript types interface file.
 //!        * ├─ pkg.js                   | The `pks.js` artifact of wasm-pack WITH bundled snippets.
 //!          ├─ pkg.js.map               | The sourcemap mapping to `pkg.js` generated by wasm-pack.
 //!        * ├─ pkg.wasm                 | The `pks_bg.wasm` artifact of wasm-pack.
-//!        * ╰─ shaders                  | Optimized shaders that contain main function code only.
-//!             ├─ list.txt              | List of optimized shaders (no extensions).
-//!             ├─ <name>.<stage>.glsl   | A single optimized shader. (Stage = vertex|fragment).
+//!        * ╰─ dynamic-assets           | Built dynamic assets.
+//!             ├─ manifest.json         | An index of all the assets and their files.
+//!             ├─ <type>                | A subdirectory for each type of asset.
+//!             │  ├─ <key>              | A subdirectory for each asset.
+//!             │  ╰─ ...
 //!             ╰─ ...
 //! ```
 //!
@@ -78,13 +78,13 @@
 //! `target/ensogl-pack/wasm-pack/index.ts`. This is the main file which when compiled glues
 //! `pkg.js`, `snippets`, and `runtime-libs.js` into a single bundle.
 //!
-//! 4. The program `target/ensogl-pack/dist/shader-extractor.js` is run. It loads
-//! `target/dist/pkg.wasm` and writes non-optimized shader code to `target/ensogl-pack/shaders`.
+//! 4. The program `target/ensogl-pack/dist/asset-extractor.js` is run. It loads
+//! `target/dist/pkg.wasm` and writes asset sources to `target/ensogl-pack/dynamic-assets`.
 //!
-//! 5. For each shader, the hash of its code is computed and compared to the hash stored in
-//! `target/ensogl-pack/shaders-hash`. If the hash did not exist or is different, the shader is
-//! optimized by using `glslc`, spirv-opt`, and `spirv-cross`, and the result is written to
-//! `dist/shaders`.
+//! 5. For each asset, its inputs are hashed and an output directory is determined based on its
+//! name and input hash. If the output directory doesn't already exist, the asset is built, and the
+//! result is written to `dist/dynamic-assets`. The manifest is rebuilt to reflect the current set
+//! of asset outputs, and any outdated output directories are removed.
 //!
 //! 6. The `target/ensogl-pack/wasm-pack/index.ts` is compiled to
 //! `target/ensogl-pack/dist/index.js`.
@@ -94,7 +94,8 @@
 //! # Runtime process.
 //! When `target/dist/index.js` is run:
 //!
-//! 1. The following files are downloaded from a server: `target/dist/{pkg.js, pkg.wasm, shaders}`.
+//! 1. The following files are downloaded from a server:
+//!    `target/dist/{pkg.js, pkg.wasm, dynamic-assets}`.
 //! 2. The code from `pkg.js` is run to compile the WASM file.
 //! 3. All before-main entry points are run.
 //! 4. Optimized shaders are uploaded to the EnsoGL application.
@@ -102,6 +103,7 @@
 
 // === Features ===
 #![feature(async_closure)]
+#![feature(fs_try_exists)]
 // === Standard Linter Configuration ===
 #![deny(non_ascii_idents)]
 #![warn(unsafe_code)]
@@ -110,19 +112,15 @@
 
 use ide_ci::prelude::*;
 
-use enso_prelude::anyhow;
 use ide_ci::program::EMPTY_ARGS;
-use ide_ci::programs::shaderc::Glslc;
-use ide_ci::programs::shaderc::SpirvOpt;
-use ide_ci::programs::spirv_cross::SpirvCross;
 use ide_ci::programs::wasm_pack::WasmPackCommand;
 use manifest_dir_macros::path;
 use std::env;
-use std::hash::Hasher;
-use std::io::ErrorKind;
 use std::path::Path;
 use std::path::PathBuf;
 use walkdir::WalkDir;
+
+mod assets;
 
 
 
@@ -438,307 +436,6 @@ async fn extract_assets(paths: &Paths) -> Result<()> {
         .await
 }
 
-/// Build the dynamic assets.
-async fn build_assets(paths: &Paths) -> Result<()> {
-    info!("Building dynamic assets.");
-    let sources = survey_asset_sources(paths)?;
-    let assets = update_assets(paths, &sources).await?;
-    let manifest = serde_json::to_string(&assets)?;
-    std::fs::write(&paths.target.ensogl_pack.dist.dynamic_assets.manifest, manifest)?;
-    gc_assets(paths, &assets)?;
-    Ok(())
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize)]
-#[serde(rename_all = "lowercase")]
-enum Builder {
-    Font,
-    Shader,
-}
-
-impl Builder {
-    fn dir_name<'a>(self) -> &'a str {
-        self.into()
-    }
-
-    async fn build_asset(
-        self,
-        input_dir: &Path,
-        input_files: &[String],
-        output_dir: &Path,
-        tmp: &Path,
-    ) -> Result<()> {
-        match self {
-            Builder::Font => build_font(input_dir, input_files, output_dir).await,
-            Builder::Shader => build_shader(input_dir, input_files, output_dir, tmp).await,
-        }
-    }
-}
-impl TryFrom<&str> for Builder {
-    type Error = anyhow::Error;
-    fn try_from(value: &str) -> std::result::Result<Self, Self::Error> {
-        match value {
-            "font" => Ok(Builder::Font),
-            "shader" => Ok(Builder::Shader),
-            other => Err(anyhow::Error::msg(format!("Unknown builder: {other:?}"))),
-        }
-    }
-}
-impl From<Builder> for &'static str {
-    fn from(value: Builder) -> Self {
-        match value {
-            Builder::Font => "font",
-            Builder::Shader => "shader",
-        }
-    }
-}
-
-struct AssetSources {
-    asset_key:   String,
-    input_files: Vec<String>,
-    inputs_hash: u64,
-}
-
-impl AssetSources {
-    fn dir_name(&self) -> String {
-        let key = &self.asset_key;
-        let hash = self.inputs_hash;
-        format!("{key}-{hash:x}")
-    }
-}
-
-#[derive(Serialize)]
-struct Asset {
-    dir:   String,
-    files: Vec<String>,
-}
-
-/// Scan the sources found in the asset sources directory.
-fn survey_asset_sources(paths: &Paths) -> Result<HashMap<Builder, Vec<AssetSources>>> {
-    let dir = ide_ci::fs::read_dir(&paths.target.ensogl_pack.dynamic_assets)?;
-    let mut asset_sources: HashMap<_, Vec<_>> = HashMap::new();
-    let mut buf = Vec::new();
-    for entry in dir {
-        let entry = entry?;
-        let builder = Builder::try_from(entry.file_name().to_string_lossy().as_ref())?;
-        let builder_dir = ide_ci::fs::read_dir(entry.path())?;
-        let builder_sources = asset_sources.entry(builder).or_default();
-        for entry in builder_dir {
-            let entry = entry?;
-            let asset_key = entry.file_name().to_string_lossy().to_string();
-            let dir = ide_ci::fs::read_dir(entry.path())?;
-            let mut file_hashes = BTreeMap::new();
-            for entry in dir {
-                let entry = entry?;
-                let file_name = entry.file_name().to_string_lossy().to_string();
-                let path = entry.path();
-                buf.clear();
-                ide_ci::fs::open(path)?.read_to_end(&mut buf)?;
-                let mut file_hasher = std::collections::hash_map::DefaultHasher::new();
-                buf.hash(&mut file_hasher);
-                file_hashes.insert(file_name, file_hasher.finish());
-            }
-            let mut asset_hasher = std::collections::hash_map::DefaultHasher::new();
-            file_hashes.hash(&mut asset_hasher);
-            let inputs_hash = asset_hasher.finish();
-            let input_files = file_hashes.into_keys().collect();
-            builder_sources.push(AssetSources { asset_key, input_files, inputs_hash });
-        }
-    }
-    Ok(asset_sources)
-}
-
-type AssetManifest = BTreeMap<Builder, BTreeMap<String, Asset>>;
-
-/// Generate any assets not found up-to-date in the cache.
-async fn update_assets(
-    paths: &Paths,
-    sources: &HashMap<Builder, Vec<AssetSources>>,
-) -> Result<AssetManifest> {
-    let out = &paths.target.ensogl_pack.dist.dynamic_assets;
-    ide_ci::fs::create_dir_if_missing(out)?;
-    let mut assets: AssetManifest = BTreeMap::new();
-    let mut deferred_assets: BTreeMap<Builder, Vec<_>> = BTreeMap::new();
-    for (&builder, builder_sources) in sources {
-        let out = out.join(builder.dir_name());
-        ide_ci::fs::create_dir_if_missing(&out)?;
-        for source_specification in builder_sources {
-            let out = out.join(source_specification.dir_name());
-            let key = source_specification.asset_key.clone();
-            match std::fs::create_dir(&out) {
-                Ok(()) => {
-                    info!("Rebuilding asset: `{}`.", out.display());
-                    let builder_assets = deferred_assets.entry(builder).or_default();
-                    let build = build_asset(paths, builder, source_specification);
-                    builder_assets.push(async move { Ok((key, build.await?)) });
-                }
-                Err(e) if e.kind() == ErrorKind::AlreadyExists => {
-                    info!("Skipping clean asset: `{}`.", out.display());
-                    let builder_assets = assets.entry(builder).or_default();
-                    let asset = survey_asset(paths, builder, source_specification)?;
-                    builder_assets.insert(key, asset);
-                }
-                Err(e) => return Err(e.into()),
-            };
-        }
-    }
-    for (builder, deferred_assets) in deferred_assets.into_iter() {
-        let deferred_assets = futures::future::join_all(deferred_assets).await;
-        let deferred_assets: Result<Vec<_>> = deferred_assets.into_iter().collect();
-        assets.entry(builder).or_default().extend(deferred_assets?);
-    }
-    Ok(assets)
-}
-
-async fn build_font(input_dir: &Path, input_files: &[String], output_dir: &Path) -> Result<()> {
-    for file_name in input_files {
-        ide_ci::fs::copy(input_dir.join(file_name), output_dir.join(file_name))?;
-    }
-    Ok(())
-}
-
-/// Generate an asset from the given sources.
-async fn build_asset(
-    paths: &Paths,
-    builder: Builder,
-    source_specification: &AssetSources,
-) -> Result<Asset> {
-    let input_dir = paths
-        .target
-        .ensogl_pack
-        .dynamic_assets
-        .join(builder.dir_name())
-        .join(&source_specification.asset_key);
-    let output_dir = paths
-        .target
-        .ensogl_pack
-        .dist
-        .dynamic_assets
-        .join(builder.dir_name())
-        .join(source_specification.dir_name());
-    let tmp_path = paths
-        .target
-        .ensogl_pack
-        .dynamic_assets
-        .join(builder.dir_name())
-        .join(format!("{}.work", source_specification.asset_key));
-    // TODO: Build the asset in a temporary directory and atomically move it into place when
-    //  complete, to prevent an interrupted build from leaving corrupted assets in the store.
-    builder
-        .build_asset(&input_dir, &source_specification.input_files, &output_dir, &tmp_path)
-        .await?;
-    survey_asset(paths, builder, source_specification)
-}
-
-/// Identify the files present in an asset directory.
-fn survey_asset(
-    paths: &Paths,
-    builder: Builder,
-    source_specification: &AssetSources,
-) -> Result<Asset> {
-    let dir = source_specification.dir_name();
-    let path = paths.target.ensogl_pack.dist.dynamic_assets.join(builder.dir_name()).join(&dir);
-    let mut files = Vec::new();
-    for entry in ide_ci::fs::read_dir(&path)? {
-        files.push(entry?.file_name().to_string_lossy().to_string());
-    }
-    Ok(Asset { dir, files })
-}
-
-/// Remove any assets not present in the manifest.
-fn gc_assets(paths: &Paths, assets: &AssetManifest) -> Result<()> {
-    let is_not_manifest = |entry: &std::io::Result<std::fs::DirEntry>| {
-        entry
-            .as_ref()
-            .map(|entry| entry.path() != paths.target.ensogl_pack.dist.dynamic_assets.manifest)
-            .unwrap_or(true)
-    };
-    for entry in paths.target.ensogl_pack.dist.dynamic_assets.read_dir()?.filter(is_not_manifest) {
-        let entry = entry?;
-        let path = entry.path();
-        let builder = Builder::try_from(entry.file_name().to_string_lossy().as_ref()).ok();
-        let assets = builder.and_then(|builder| assets.get(&builder));
-        match assets {
-            Some(assets) => {
-                let assets: HashSet<_> = assets.values().map(|asset| asset.dir.as_ref()).collect();
-                for entry in path.read_dir()? {
-                    let entry = entry?;
-                    let path = entry.path();
-                    if !assets.contains(entry.file_name().to_string_lossy().as_ref()) {
-                        info!("Cleaning unused asset at `{}`.", path.display());
-                        ide_ci::fs::remove_if_exists(path)?;
-                    }
-                }
-            }
-            _ => {
-                info!("Cleaning unused builder at `{}`.", path.display());
-                ide_ci::fs::remove_if_exists(path)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Build optimized shaders by using `glslc`, `spirv-opt` and `spirv-cross`.
-async fn build_shader(
-    input_dir: &Path,
-    input_files: &[String],
-    output_dir: &Path,
-    work_dir: &Path,
-) -> Result<()> {
-    ide_ci::fs::create_dir_if_missing(work_dir)?;
-    info!("Optimizing `{}`.", input_dir.file_name().unwrap().to_string_lossy());
-    for glsl_file_name in input_files {
-        let glsl_path = input_dir.join(glsl_file_name);
-        let work_path = work_dir.join(glsl_file_name);
-        let stage_path = work_path.with_extension("");
-        let stage = stage_path.file_name().unwrap().to_string_lossy();
-        let spv_path = stage_path.with_appended_extension("spv");
-        let spv_opt_path = stage_path.with_appended_extension("opt.spv");
-        let glsl_opt_path = stage_path.with_appended_extension("opt.glsl");
-        let glsl_opt_dist_path = output_dir.join(&glsl_file_name);
-        let spv_path = spv_path.as_str();
-        let glsl_path = glsl_path.as_str();
-        let shader_stage = &format!("-fshader-stage={stage}");
-        let glslc_args = ["--target-env=opengl", shader_stage, "-o", spv_path, glsl_path];
-        let spirv_opt_args = ["-O", "-o", spv_opt_path.as_str(), spv_path.as_str()];
-        let spirv_cross_args = ["--output", glsl_opt_path.as_str(), spv_opt_path.as_str()];
-        Glslc.cmd()?.args(glslc_args).run_ok().await?;
-        SpirvOpt.cmd()?.args(spirv_opt_args).run_ok().await?;
-        SpirvCross.cmd()?.args(spirv_cross_args).run_ok().await?;
-
-        let content = ide_ci::fs::read_to_string(&glsl_opt_path)?.replace("\r\n", "\n");
-        let extract_err = || format!("Failed to process shader '{}'.", glsl_opt_path.as_str());
-        let code = extract_main_shader_code(&content).with_context(extract_err)?;
-        ide_ci::fs::write(&glsl_opt_dist_path, code)?;
-    }
-    Ok(())
-}
-
-/// Read the optimized shader code, extract the main function body and preserve all top-level
-/// variable declarations.
-fn extract_main_shader_code(code: &str) -> Result<String> {
-    let main_start_str = "void main()\n{";
-    let main_end_str = "}";
-    let main_fn_find_err = "Failed to find main function.";
-    let main_start = code.find(main_start_str).with_context(|| main_fn_find_err)?;
-    let main_end = code.rfind(main_end_str).with_context(|| main_fn_find_err)?;
-    let before_main = &code[..main_start];
-    let declarations: Vec<&str> = before_main
-        .lines()
-        .filter_map(|line| {
-            let version_def = line.starts_with("#version ");
-            let precision_def = line.starts_with("precision ");
-            let layout_def = line.starts_with("layout(");
-            let def = version_def || precision_def || layout_def;
-            (!def).then_some(line)
-        })
-        .collect();
-    let declarations = declarations.join("\n");
-    let main_content = &code[main_start + main_start_str.len()..main_end];
-    Ok(format!("{declarations}\n{main_content}"))
-}
-
 /// Wrapper over `wasm-pack build` command.
 ///
 /// # Arguments
@@ -753,7 +450,7 @@ pub async fn build(
     compile_this_crate_ts_sources(&paths).await?;
     run_wasm_pack(&paths, provider).await?;
     extract_assets(&paths).await?;
-    build_assets(&paths).await?;
+    assets::build(&paths).await?;
     let out_dir = Path::new(&outputs.out_dir);
     ide_ci::fs::copy(&paths.target.ensogl_pack.dist, out_dir)?;
     ide_ci::fs::remove_symlink_dir_if_exists(&paths.target.ensogl_pack.linked_dist)?;
